--- src/main/java/net/ciebus/pinmod/CommonProxy.java
@@ -1,5 +1,0 @@
-package net.ciebus.pinmod;
-
-public class CommonProxy {
-    public void registerClientInfo(){}
-}
