+++ src/main/java/net/ciebus/pinmod/Pin.java
@@ -1,54 +1,29 @@
 package net.ciebus.pinmod;
 
-import cpw.mods.fml.common.FMLCommonHandler;
-import cpw.mods.fml.common.SidedProxy;
-import cpw.mods.fml.common.event.FMLPreInitializationEvent;
-import cpw.mods.fml.common.eventhandler.SubscribeEvent;
-import cpw.mods.fml.common.gameevent.InputEvent;
-import net.minecraft.client.Minecraft;
-import net.minecraft.init.Blocks;
 import cpw.mods.fml.common.Mod;
 import cpw.mods.fml.common.Mod.EventHandler;
+import cpw.mods.fml.common.SidedProxy;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
-import net.minecraft.util.MovingObjectPosition;
-import net.minecraftforge.common.MinecraftForge;
+import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+import net.ciebus.pinmod.common.CommonProxy;
 
 @Mod(modid = Pin.MODID, version = Pin.VERSION)
-public class Pin {
+public final class Pin {
+
     public static final String MODID = "pin";
     public static final String VERSION = "1.0";
 
-    @SidedProxy(clientSide = "net.ciebus.pinmod.ClientProxy", serverSide = "net.ciebus.pinmod.CommonProxy")
+    @SidedProxy(clientSide = "net.ciebus.pinmod.client.ClientProxy", serverSide = "net.ciebus.pinmod.common.CommonProxy")
     public static CommonProxy proxy;
 
     @EventHandler
-    public void init(FMLInitializationEvent event) {
+    public static void preInit(FMLPreInitializationEvent event) {
+        proxy.onPreInit(event);
     }
-
 
     @EventHandler
-    public void load(FMLInitializationEvent event) {
-        FMLCommonHandler.instance().bus().register(this);
-        FMLCommonHandler.instance().bus().register(new PinManager());
-        MinecraftForge.EVENT_BUS.register(new PinManager());
-        proxy.registerClientInfo();
+    public static void init(FMLInitializationEvent event) {
+        proxy.onInit(event);
     }
 
-
-    @EventHandler
-    public void preInit(FMLPreInitializationEvent event) {
-        PacketHandler.init();
-    }
-
-
-    @SubscribeEvent
-    public void KeyHandlingEvent(InputEvent event) {
-        if (ClientProxy.sampleKey.isPressed()) {
-            MovingObjectPosition mop = Minecraft.getMinecraft().renderViewEntity.rayTrace(200, 1.0F);
-            if (mop != null) {
-                boolean state = PinManager.isDelete(mop.hitVec.xCoord, mop.hitVec.yCoord, mop.hitVec.zCoord, Minecraft.getMinecraft().thePlayer.getDisplayName(), Minecraft.getMinecraft().thePlayer.worldObj.provider.dimensionId);
-                PacketHandler.INSTANCE.sendToServer(new MessageKeyPressed(state, mop.hitVec.xCoord, mop.hitVec.yCoord, mop.hitVec.zCoord, Minecraft.getMinecraft().thePlayer.getDisplayName(), Minecraft.getMinecraft().thePlayer.worldObj.provider.dimensionId));
-            }
-        }
-    }
 }
