+++ src/main/java/net/ciebus/pinmod/common/CommonProxy.java
@@ -1,0 +1,16 @@
+package net.ciebus.pinmod.common;
+
+import cpw.mods.fml.common.event.FMLInitializationEvent;
+import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+import net.ciebus.pinmod.common.network.PacketHandler;
+
+public class CommonProxy {
+
+    public void onPreInit(FMLPreInitializationEvent event) {
+        PacketHandler.init();
+    }
+
+    public void onInit(FMLInitializationEvent event) {
+    }
+
+}
