+++ src/main/java/net/ciebus/pinmod/common/network/CMessageKeyPressedHandler.java
@@ -1,8 +1,9 @@
-package net.ciebus.pinmod;
+package net.ciebus.pinmod.common.network;
 
 import cpw.mods.fml.common.network.simpleimpl.IMessage;
 import cpw.mods.fml.common.network.simpleimpl.IMessageHandler;
 import cpw.mods.fml.common.network.simpleimpl.MessageContext;
+import net.ciebus.pinmod.PinManager;
 
 public class CMessageKeyPressedHandler implements IMessageHandler<MessageKeyPressed, IMessage> {
