+++ src/main/java/net/ciebus/pinmod/common/network/SMessageKeyPressedHandler.java
@@ -1,11 +1,8 @@
-package net.ciebus.pinmod;
+package net.ciebus.pinmod.common.network;
 
 import cpw.mods.fml.common.network.simpleimpl.IMessage;
 import cpw.mods.fml.common.network.simpleimpl.IMessageHandler;
 import cpw.mods.fml.common.network.simpleimpl.MessageContext;
-import net.minecraft.client.Minecraft;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.util.ChatComponentText;
 
 public class SMessageKeyPressedHandler implements IMessageHandler<MessageKeyPressed, IMessage> {
 
@@ -14,4 +11,5 @@ public class SMessageKeyPressedHandler implements IMessageHandler<MessageKeyPres
         PacketHandler.INSTANCE.sendToAll(message);
         return null;
     }
+
 }
